+++ examples/src/lightweight.rs
@@ -4,7 +4,7 @@ use autd3_protobuf::lightweight::LightweightClient;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    let mut client = LightweightClient::builder([AUTD3::new(Vector3::zeros())])
+    let mut client = LightweightClient::builder([AUTD3::new(Point3::origin())])
         .open("127.0.0.1:8080".parse()?)
         .await?;
 
@@ -19,9 +19,7 @@ async fn main() -> anyhow::Result<()> {
     println!("============================================");
 
     client.send(Sine::new(150. * Hz)).await?;
-    client
-        .send(Focus::new(Vector3::new(90., 70., 150.)))
-        .await?;
+    client.send(Focus::new(Point3::new(90., 70., 150.))).await?;
 
     println!("Press enter to exit...");
     let mut _s = String::new();